+++ src/share/ioModule.f90
@@ -5,6 +5,7 @@ module ioModule
   use runInfoType
   use forcingType
   use modelVarType
+  use derivedType
   
   implicit none
   
@@ -294,7 +295,7 @@ contains
       write(*,'("Problem opening file ''", A, "''")') trim(filename)
       stop ":  ERROR EXIT"
     endif
-    write(runinfo%output_fileunits(1),'(A)') 'year mo dy hr tair precip pet qs qg tci eta'   ! header
+    write(runinfo%output_fileunits(1),'(A)') 'year mo dy hr tair precip pet eta qs qg tci roimp sdro ssur sif bfs bfp'   ! header
 
     ! if user setting is to write out information for each snowband, open the individual files
     if (namelist%output_hrus == 1) then
@@ -311,7 +312,7 @@ contains
         endif
       
         ! Write 1-line header
-        write(runinfo%output_fileunits(nh+1),'(A)') 'year mo dy hr tair precip pet qs qg tci eta'
+        write(runinfo%output_fileunits(nh+1),'(A)') 'year mo dy hr tair precip pet eta qs qg tci roimp sdro ssur sif bfs bfp'
         
       end do  ! end loop over sub-units
       
@@ -462,7 +463,7 @@ contains
 
   ! === write output for one timestep ===
   ! assumes that files have been opened and header already written
-  SUBROUTINE write_sac_output(namelist, runinfo, parameters, forcing, modelvar, n_curr_hru)
+  SUBROUTINE write_sac_output(namelist, runinfo, parameters, forcing, modelvar, derived, n_curr_hru)
     implicit none
     type (namelist_type),    intent(in)     :: namelist
     type (runinfo_type),     intent(in)     :: runinfo
@@ -472,6 +473,7 @@ contains
     ! forcing & modelvar are because the combined variables get updated here
     type (forcing_type),     intent(inout)  :: forcing
     type (modelvar_type),    intent(inout)  :: modelvar
+    type (derived_type),     intent(inout)  :: derived
     
     ! local variables
     integer        :: ierr   ! error code returned by open(iostat = ierr)
@@ -480,13 +482,15 @@ contains
     ! ==== WRITE output for current area simulation ====
     ! Note:  write order should match header written by open_and_init_output_files()
     
-    32 FORMAT(I4.4, 3(1x,I2.2), 6(F10.3))
+    32 FORMAT(I4.4, 3(1x,I2.2), 20(F10.3))
 
     ! if user setting is to write out information for each snowband, open the individual files
     if (namelist%output_hrus == 1 .and. runinfo%n_hrus > 1) then
       write(runinfo%output_fileunits(n_curr_hru+1), 32, iostat=ierr) runinfo%curr_yr, runinfo%curr_mo, runinfo%curr_dy, runinfo%curr_hr, &
-            forcing%tair(n_curr_hru), forcing%precip(n_curr_hru), forcing%pet(n_curr_hru), &
-            modelvar%qs(n_curr_hru), modelvar%qg(n_curr_hru), modelvar%tci(n_curr_hru), modelvar%eta(n_curr_hru)
+            forcing%tair(n_curr_hru), forcing%precip(n_curr_hru), forcing%pet(n_curr_hru), modelvar%eta(n_curr_hru), &
+            modelvar%qs(n_curr_hru), modelvar%qg(n_curr_hru), modelvar%tci(n_curr_hru), & 
+            modelvar%roimp(n_curr_hru), modelvar%sdro(n_curr_hru), modelvar%ssur(n_curr_hru), &
+            modelvar%sif(n_curr_hru), modelvar%bfs(n_curr_hru), modelvar%bfp(n_curr_hru)
       if(ierr /= 0) then
         print*, 'ERROR writing output information for basin average'; stop
       endif            
@@ -495,38 +499,58 @@ contains
     ! ==== if all hrus have been run, sum across hrus with weighting for snowband area ====
     
     ! initialize for timestep
-    forcing%tair_comb       = 0.0
-    forcing%precip_comb     = 0.0
-    forcing%pet_comb        = 0.0
-    modelvar%qs_comb        = 0.0
-    modelvar%qg_comb        = 0.0
-    modelvar%tci_comb       = 0.0
-    modelvar%eta_comb       = 0.0
+    derived%tair_comb       = 0.0
+    derived%precip_comb     = 0.0
+    derived%pet_comb        = 0.0
+    derived%qs_comb        = 0.0
+    derived%qg_comb        = 0.0
+    derived%tci_comb       = 0.0
+    derived%eta_comb       = 0.0
+    derived%roimp_comb     = 0.0
+    derived%sdro_comb      = 0.0
+    derived%ssur_comb      = 0.0
+    derived%sif_comb       = 0.0
+    derived%bfs_comb       = 0.0
+    derived%bfp_comb       = 0.0
         
     if (n_curr_hru .eq. runinfo%n_hrus) then 
       do nh=1, runinfo%n_hrus
-        forcing%tair_comb        = forcing%tair_comb + forcing%tair(nh) * parameters%hru_area(nh)
-        forcing%precip_comb      = forcing%precip_comb + forcing%precip(nh) * parameters%hru_area(nh)
-        forcing%pet_comb         = forcing%pet_comb + forcing%pet(nh) * parameters%hru_area(nh)
-        modelvar%qs_comb         = modelvar%qs_comb + modelvar%qs(nh) * parameters%hru_area(nh) 
-        modelvar%qg_comb         = modelvar%qg_comb + modelvar%qg(nh) * parameters%hru_area(nh) 
-        modelvar%tci_comb        = modelvar%tci_comb + modelvar%tci(nh) * parameters%hru_area(nh)
-        modelvar%eta_comb        = modelvar%eta_comb + modelvar%eta(nh) * parameters%hru_area(nh)
+        derived%tair_comb        = derived%tair_comb + forcing%tair(nh) * parameters%hru_area(nh)
+        derived%precip_comb      = derived%precip_comb + forcing%precip(nh) * parameters%hru_area(nh)
+        derived%pet_comb         = derived%pet_comb + forcing%pet(nh) * parameters%hru_area(nh)
+        derived%qs_comb          = derived%qs_comb + modelvar%qs(nh) * parameters%hru_area(nh) 
+        derived%qg_comb          = derived%qg_comb + modelvar%qg(nh) * parameters%hru_area(nh) 
+        derived%tci_comb         = derived%tci_comb + modelvar%tci(nh) * parameters%hru_area(nh)
+        derived%eta_comb         = derived%eta_comb + modelvar%eta(nh) * parameters%hru_area(nh)
+        derived%roimp_comb       = derived%roimp_comb + modelvar%roimp(nh) * parameters%hru_area(nh)
+        derived%sdro_comb        = derived%sdro_comb  + modelvar%sdro(nh) * parameters%hru_area(nh)
+        derived%ssur_comb        = derived%ssur_comb  + modelvar%ssur(nh) * parameters%hru_area(nh)
+        derived%sif_comb         = derived%sif_comb + modelvar%sif(nh) * parameters%hru_area(nh)
+        derived%bfs_comb         = derived%bfs_comb + modelvar%bfs(nh) * parameters%hru_area(nh)
+        derived%bfp_comb         = derived%bfp_comb + modelvar%bfp(nh) * parameters%hru_area(nh)
       end do
 
       ! take average of weighted sum of HRU areas
-      forcing%tair_comb        = forcing%tair_comb / parameters%total_area
-      forcing%precip_comb      = forcing%precip_comb / parameters%total_area
-      forcing%pet_comb         = forcing%pet_comb / parameters%total_area
-      modelvar%qs_comb         = modelvar%qs_comb / parameters%total_area
-      modelvar%qg_comb         = modelvar%qg_comb / parameters%total_area
-      modelvar%tci_comb        = modelvar%tci_comb / parameters%total_area
-      modelvar%eta_comb        = modelvar%eta_comb / parameters%total_area
+      derived%tair_comb        = derived%tair_comb / parameters%total_area
+      derived%precip_comb      = derived%precip_comb / parameters%total_area
+      derived%pet_comb         = derived%pet_comb / parameters%total_area
+      derived%qs_comb          = derived%qs_comb / parameters%total_area
+      derived%qg_comb          = derived%qg_comb / parameters%total_area
+      derived%tci_comb         = derived%tci_comb / parameters%total_area
+      derived%eta_comb         = derived%eta_comb / parameters%total_area
+      derived%roimp_comb       = derived%roimp_comb / parameters%total_area
+      derived%sdro_comb        = derived%sdro_comb / parameters%total_area
+      derived%ssur_comb        = derived%ssur_comb / parameters%total_area
+      derived%sif_comb         = derived%sif_comb / parameters%total_area
+      derived%bfs_comb         = derived%bfs_comb / parameters%total_area
+      derived%bfp_comb         = derived%bfp_comb / parameters%total_area
 
       ! -- write out combined file that is similar to each area file, but add flow variable in CFS units
       write(runinfo%output_fileunits(1), 32, iostat=ierr) runinfo%curr_yr, runinfo%curr_mo, runinfo%curr_dy, runinfo%curr_hr, &
-            forcing%tair_comb, forcing%precip_comb, forcing%pet_comb, &
-            modelvar%qs_comb, modelvar%qg_comb, modelvar%tci_comb, modelvar%eta_comb
+            derived%tair_comb, derived%precip_comb, derived%pet_comb, derived%eta_comb, &
+            derived%qs_comb, derived%qg_comb, derived%tci_comb, &
+            derived%roimp_comb, derived%sdro_comb, derived%ssur_comb, &
+            derived%sif_comb, derived%bfs_comb, derived%bfp_comb
       if(ierr /= 0) then
         print*, 'ERROR writing output information for sub-unit ', n_curr_hru; stop
       endif